refactor writers into their own directory
