+++ unstructured/ingest/connector/fsspec.py
@@ -2,7 +2,7 @@ import json
 import os
 import typing as t
 from contextlib import suppress
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from pathlib import Path, PurePath
 
 from unstructured.ingest.compression_support import (
@@ -20,6 +20,7 @@ from unstructured.ingest.interfaces import (
     IngestDocCleanupMixin,
     SourceConnectorCleanupMixin,
     SourceMetadata,
+    WriteConfig,
 )
 from unstructured.ingest.logger import logger
 from unstructured.utils import (
@@ -222,8 +223,14 @@ class FsspecSourceConnector(
 
 
 @dataclass
+class FsspecWriteConfig(WriteConfig):
+    write_text_kwargs: t.Dict[str, t.Any] = field(default_factory=dict)
+
+
+@dataclass
 class FsspecDestinationConnector(BaseDestinationConnector):
     connector_config: SimpleFsspecConfig
+    write_config: FsspecWriteConfig
 
     def initialize(self):
         from fsspec import AbstractFileSystem, get_filesystem_class
@@ -249,15 +256,16 @@ class FsspecDestinationConnector(BaseDestinationConnector):
 
         logger.info(f"Writing content using filesystem: {type(fs).__name__}")
 
-        output_folder = self.connector_config.path_without_protocol
-        output_folder = os.path.join(output_folder)  # Make sure folder ends with file seperator
-        filename = (
-            filename.strip(os.sep) if filename else filename
-        )  # Make sure filename doesn't begin with file seperator
-        output_path = str(PurePath(output_folder, filename)) if filename else output_folder
-        full_output_path = f"s3://{output_path}"
-        logger.debug(f"uploading content to {full_output_path}")
-        fs.write_text(full_output_path, json.dumps(json_list, indent=indent), encoding=encoding)
+        dest_folder = self.connector_config.path_without_protocol
+        dest_output_path = str(PurePath(dest_folder, filename)) if filename else dest_folder
+        full_dest_path = f"{self.connector_config.protocol}://{dest_output_path}"
+        logger.debug(f"uploading content to {full_dest_path}")
+        fs.write_text(
+            full_dest_path,
+            json.dumps(json_list, indent=indent),
+            encoding=encoding,
+            **self.write_config.write_text_kwargs,
+        )
 
     def write(self, docs: t.List[BaseIngestDoc]) -> None:
         for doc in docs:
