+++ unstructured/ingest/runner/writers/delta_table.py
@@ -1,0 +1,22 @@
+import typing as t
+from pathlib import Path
+
+
+def delta_table_writer(
+    table_uri: t.Union[str, Path],
+    write_column: str,
+    mode: t.Literal["error", "append", "overwrite", "ignore"] = "error",
+    **kwargs,
+):
+    from unstructured.ingest.connector.delta_table import (
+        DeltaTableDestinationConnector,
+        DeltaTableWriteConfig,
+        SimpleDeltaTableConfig,
+    )
+
+    return DeltaTableDestinationConnector(
+        write_config=DeltaTableWriteConfig(write_column=write_column, mode=mode),
+        connector_config=SimpleDeltaTableConfig(
+            table_uri=table_uri,
+        ),
+    )
